+++ src/cjs/VEllipseSector.js
@@ -9,8 +9,11 @@
  */
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.VEllipseSector = void 0;
+// TODO: add class to all demos
+// TODO: add to PlotBoilerplate.add(...)
 var VEllipse_1 = require("./VEllipse");
 var UIDGenerator_1 = require("./UIDGenerator");
+var Vertex_1 = require("./Vertex");
 /**
  * @classdesc A class for elliptic sectors.
  *
@@ -54,12 +57,16 @@ var VEllipseSector = /** @class */ (function () {
          * @param {boolean} options.moveToStart - If false (default=true) the initial 'Move' command will not be used.
          * @return [ 'A', radiusH, radiusV, rotation=0, largeArcFlag=1|0, sweepFlag=0, endx, endy ]
          */
-        describeSVGArc: function (x, y, radiusH, radiusV, startAngle, endAngle, options) {
+        describeSVGArc: function (x, y, radiusH, radiusV, startAngle, endAngle, rotation, options) {
             if (typeof options === 'undefined')
                 options = { moveToStart: true };
+            if (typeof rotation === 'undefined')
+                rotation = 0.0;
             // XYCoords
-            var end = VEllipse_1.VEllipse.utils.polarToCartesian(x, y, radiusH, radiusV, endAngle);
-            var start = VEllipse_1.VEllipse.utils.polarToCartesian(x, y, radiusH, radiusV, startAngle);
+            var end = new Vertex_1.Vertex(VEllipse_1.VEllipse.utils.polarToCartesian(x, y, radiusH, radiusV, endAngle));
+            var start = new Vertex_1.Vertex(VEllipse_1.VEllipse.utils.polarToCartesian(x, y, radiusH, radiusV, startAngle));
+            end.rotate(rotation, { x: x, y: y });
+            start.rotate(rotation, { x: x, y: y });
             var diff = endAngle - startAngle;
             /*
               var r2d = 180/Math.PI;
@@ -83,7 +90,7 @@ var VEllipseSector = /** @class */ (function () {
             if (options.moveToStart) {
                 pathData.push('M', start.x, start.y);
             }
-            pathData.push("A", radiusH, radiusV, 0, largeArcFlag, sweepFlag, end.x, end.y);
+            pathData.push("A", radiusH, radiusV, -rotation, largeArcFlag, sweepFlag, end.x, end.y);
             return pathData;
         } // END function describeSVGArc
     }; // END ellipseSectorUtils