+++ src/cjs/draw.d.ts
@@ -293,11 +293,12 @@ export declare class drawutils implements DrawLib<void> {
      * @param {number} radiusY - The radius of the ellipse.
      * @param {string} color - The CSS color to draw the ellipse with.
      * @param {number} lineWidth=1 - An optional line width param (default is 1).
+     * @param {number=} rotation - (optional, default=0) The rotation of the ellipse.
      * @return {void}
      * @instance
      * @memberof drawutils
      */
-    ellipse(center: Vertex, radiusX: number, radiusY: number, color: string, lineWidth?: number): void;
+    ellipse(center: Vertex, radiusX: number, radiusY: number, color: string, lineWidth?: number, rotation?: number): void;
     /**
      * Draw square at the given center, size and with the specified (CSS-) color.<br>
      * <br>