+++ src/cjs/drawgl.js
@@ -343,11 +343,12 @@ var drawutilsgl = /** @class */ (function () {
      * @param {number} radiusY - The radius of the ellipse.
      * @param {string} color - The CSS color to draw the ellipse with.
      * @param {number=} lineWidth - (optional) The line width to use; default is 1.
+     * @param {number=} rotation - (optional, default=0) The rotation of the ellipse.
      * @return {void}
      * @instance
      * @memberof drawutils
      */
-    drawutilsgl.prototype.ellipse = function (center, radiusX, radiusY, color, lineWidth) {
+    drawutilsgl.prototype.ellipse = function (center, radiusX, radiusY, color, lineWidth, rotation) {
         // NOT YET IMPLEMENTED
     };
     ;