+++ src/ts/drawgl.ts
@@ -419,11 +419,12 @@ export class drawutilsgl implements DrawLib<void> {
      * @param {number} radiusY - The radius of the ellipse.
      * @param {string} color - The CSS color to draw the ellipse with.
      * @param {number=} lineWidth - (optional) The line width to use; default is 1.
+     * @param {number=} rotation - (optional, default=0) The rotation of the ellipse.
      * @return {void}
      * @instance
      * @memberof drawutils
      */
-    ellipse( center:Vertex, radiusX:number, radiusY:number, color:string, lineWidth?:number) {
+    ellipse( center:Vertex, radiusX:number, radiusY:number, color:string, lineWidth?:number, rotation?:number ) {
 	// NOT YET IMPLEMENTED
     };