+++ src/ts/utils/helpers/drawutilssvg.ts
@@ -649,19 +649,25 @@ export class drawutilssvg implements DrawLib<void|SVGElement> {
      * @param {number} radiusY - The radius of the ellipse.
      * @param {string} color - The CSS color to draw the ellipse with.
      * @param {number=} lineWidth - (optional) The line width to use; default is 1.
+     * @param {number=} rotation - (optional, default=0) The rotation of the ellipse.
      * @return {void}
      * @instance
      * @memberof drawutilssvg
      */
-    ellipse( center:Vertex, radiusX:number, radiusY:number, color:string, lineWidth?:number ) {
+    ellipse( center:Vertex, radiusX:number, radiusY:number, color:string, lineWidth?:number, rotation?:number ) {
+	if( typeof rotation === 'undefined' ) {
+	    rotation = 0.0;
+	}
 	const node : SVGElement = this.makeNode('ellipse');	
 	node.setAttribute( 'cx', `${this._x(center.x)}` );
 	node.setAttribute( 'cy', `${this._y(center.y)}` );
 	node.setAttribute( 'rx', `${radiusX * this.scale.x}` );
 	node.setAttribute( 'ry', `${radiusY * this.scale.y}` );
+	// node.setAttribute( 'style', `transform: rotate(${rotation} ${center.x} ${center.y})` );
+	node.setAttribute( 'transform', `rotate(${rotation*180/Math.PI} ${this._x(center.x)} ${this._y(center.y)})` );
 	
 	return this._bindFillDraw( node, 'ellipse', color, lineWidth || 1 );
-    };   
+    };
 
 
     /**